banyan-polars/pyproject.toml
[tool.poetry]
name = "banyan-polars"
version = "0.1.0"
description = ""
authors = ["cailinw <cailinwin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
polars = "^0.13.58"
banyan-python = "^0.2.0"
mpi4py = "^3.1.3"
plum-dispatch = "^1.6"
pyarrow = "^8.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
